requirements.txt
# Core
torch>=2.0
torchvision>=0.15

# CLIP model, tokenizer, and pretrained weights
open-clip-torch>=2.20

# PyTorch Lightning (training loop, DDP, callbacks)
lightning>=2.0

# Hydra + OmegaConf (config management)
hydra-core>=1.3
omegaconf>=2.3

# Vision models (Swin, MobileViT, ResNet-18 via timm backend in open_clip)
timm>=0.9

# HuggingFace (cosine schedule, optional HF text encoders)
transformers>=4.30

# Data loading
pandas
webdataset>=0.2.5

# Image processing
Pillow

# Utilities
ftfy
regex
tqdm
numpy
